[lldb] Extract Mach-O export trie parsing into MachOTrie.{h,cpp} (NFC)
ParseTrieEntries and its TrieEntry helpers were file-local statics in
ObjectFileMachO.cpp, unreachable from tests. Move them into a
self-contained translation unit (depending only on lldbUtility) so the
parser can be exercised in isolation.
