+++ lldb/source/Plugins/ObjectFile/Mach-O/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_lldb_library(lldbPluginObjectFileMachO PLUGIN
   ObjectFileMachO.cpp
+  MachOTrie.cpp
 
   LINK_COMPONENTS
     Support
